Cast floats

+++ src/single.cc
@@ -86,14 +86,14 @@ int main (int argc, char* argv[]) {
 		graphics_window.clear();
 		
 		// Render sprites to screen
-		ball.setPosition(screen.x(0) - screen.radius(0), screen.y(0) - screen.radius(0));
-		dial.setRotation(screen.velocity(0).dir + 90);
+		ball.setPosition(static_cast<float>(screen.x(0) - screen.radius(0)), static_cast<float>(screen.y(0) - screen.radius(0)));
+		dial.setRotation(static_cast<float>(screen.velocity(0).dir + 90));
 		graphics_window.draw(ball);
 		graphics_window.draw(dial);
 
 		graphics_window.draw(mag);
 
-		mag_inner.setScale(screen.velocity(0).mag / max_speed, 1);
+		mag_inner.setScale(static_cast<float>(screen.velocity(0).mag / max_speed), 1);
 		graphics_window.draw(mag_inner);
 		std::cout << "\rVelocity: " << screen.velocity(0).mag << "\t\tDirection: " << screen.velocity(0).dir << "     ";
 		graphics_window.display(); // Update screen based on changes
